ENG-4087 add warning message on cloning bundle without descriptor json

+++ test/commands/mfe/rm.test.ts
@@ -83,9 +83,11 @@ describe('mfe rm', () => {
     process.chdir(tempBundleDir)
     bundleDescriptorService = new BundleDescriptorService()
     bundleDescriptorService.writeBundleDescriptor(bundleDescriptor)
-    fs.rmdirSync(path.resolve(tempBundleDir, ...OUTPUT_FOLDER), {
-      recursive: true
-    })
+    if (fs.existsSync(path.resolve(tempBundleDir, ...OUTPUT_FOLDER))) {
+      fs.rmdirSync(path.resolve(tempBundleDir, ...OUTPUT_FOLDER), {
+        recursive: true
+      })
+    }
   })
 
   afterEach(() => {
